+++ packages/base/Input/src/Input/styles.ts
@@ -1,16 +1,12 @@
-import type { Theme } from '@material-ui/core/styles'
 import { createStyles } from '@material-ui/core/styles'
 import '@toptal/picasso-outlined-input/styles'
 import '@toptal/picasso-input-adornment/styles'
 
-import highlightAutofillStyles from '../InputBase/highlight-styles'
 import '../InputBase/styles'
 
-export default (theme: Theme) =>
+export default () =>
   createStyles({
     root: {
-      fontSize: '1rem',
-      backgroundColor: theme.palette.common.white,
       cursor: 'text',
     },
     rootMultiline: {
@@ -19,8 +15,4 @@ export default (theme: Theme) =>
     inputMultilineResizable: {
       resize: 'vertical',
     },
-    horizontalLayout: {
-      width: '100%',
-    },
-    ...highlightAutofillStyles(theme),
   })