+++ packages/base/Input/src/index.ts
@@ -1,2 +1,1 @@
 export * from './Input'
-export { default as highlightStyles } from './InputBase/highlight-styles'