+++ packages/base/Select/src/NativeSelect/styles.ts
@@ -3,8 +3,6 @@ import type { Theme } from '@material-ui/core/styles'
 import { createStyles } from '@material-ui/core/styles'
 import '@toptal/picasso-input/styles'
 
-import { highlightStyles as highlightAutofillStyles } from '@toptal/picasso-input'
-
 export default (theme: Theme) => {
   const { palette } = theme
 
@@ -54,6 +52,5 @@ export default (theme: Theme) => {
     endAdornmentPadding: {
       paddingRight: '3.5625rem',
     },
-    ...highlightAutofillStyles(theme),
   })
 }